requirements.txt
python-dotenv==1.0.0
openai==1.12.0
python-docx==0.8.11